float error

+++ WP90_1.c
@@ -1,10 +1,10 @@
 #include <stdio.h> //*sqrt() don't use math.h
 int main()
 {
-    int num;
+    float num;
     float x[10000], result = 0;
     printf("Enter number : ");
-    scanf("%d", &num);
+    scanf("%f", &num);
     for (int i = 0; i <= 10000; i++)
     {
         x[0] = 1;
